# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Culbridge Compliance API",
    "description": "Deterministic compliance engine for agricultural export compliance\n\n## Key Features\n- Deterministic rule evaluation (no probabilistic outcomes)\n- Full audit trail for regulatory compliance\n- Multi-corridor support (EU, USA, etc.)\n- Lab and document verification\n- HS Code validation\n\n## Error Handling\nAll endpoints return standardized error responses. See `EngineError` schema.",
    "version": "1.0.0",
    "contact": {
      "name": "Culbridge API Support",
      "email": "support@culbridge.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.culbridge.com/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                }
              },
              "example": {
                "email": "exporter@company.com",
                "password": "securepassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EngineError" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "JWT access token" },
          "refresh_token": { "type": "string" },
          "expires_in": { "type": "integer", "description": "Seconds" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "expires_in": { "type": "integer" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["exporter", "compliance", "admin", "founder"] },
          "name": { "type": "string" },
          "entity_id": { "type": "string", "description": "Associated exporter/company ID" }
        }
      },
...
      "EngineError": {
        "type": "object",
        "description": "Standardized error object for all engine errors",
        "properties": {
          "name": { "type": "string", "description": "Error class name" },
          "severity": { "type": "string", "enum": ["BLOCKER", "WARNING", "INFO"], "description": "Error severity level" },
          "code": { "type": "string", "description": "Unique error code (e.g., API_TIMEOUT, MISSING_DOCUMENT)" },
          "message": { "type": "string", "description": "Human-readable message" },
          "context": { "type": "object", "description": "Additional context (shipmentId, labId, etc.)" },
          "timestamp": { "type": "string", "format": "date-time" },
          "stack": { "type": "string", "description": "Error stack trace (dev only)" }
        },
        "required": ["severity", "code", "message", "timestamp"]
      },
...
    }
  }
}